moved html structure for companlist and movielist to another component. Moved html structure from MoviePage to another component
