+++ code/src/components/Nav.js
@@ -8,16 +8,12 @@ const Nav = () => (
 		<Link to="/">
 			<h1>Movie time</h1>
 		</Link>
-
 		<nav>
 			<button className="nav-button">Category</button>
 			<div className="nav-links">
 				<NavLink to="/">Now Playing</NavLink>
-
 				<NavLink to="/popular">Popular</NavLink>
-
 				<NavLink to="/upcoming">Upcoming</NavLink>
-
 				<NavLink to="/top_rated">Top Rated</NavLink>
 			</div>
 		</nav>
